+++ lib/extensions/email_confirmation/phoenix/controllers/controller_callbacks.ex
@@ -4,22 +4,14 @@ defmodule PowEmailConfirmation.Phoenix.ControllerCallbacks do
 
   ### User hasn't confirmed e-mail
 
-  Triggers on `Pow.Phoenix.RegistrationController.create/2` or
+  Triggers on `Pow.Phoenix.RegistrationController.create/2` and
   `Pow.Phoenix.SessionController.create/2`.
 
   When a user is created or authenticated, and the current e-mail hasn't been
-  confirmed, a confirmation e-mail is sent, the session will be cleared, and the
-  user redirected back to `Pow.Phoenix.Routes.after_registration_path/1` or
-  `Pow.Phoenix.Routes.after_registration_path/1` respectively.
-
-  ### E-mail already taken during registration
-
-  Triggers on `Pow.Phoenix.RegistrationController.create/2`.
-
-  When an error is returned during registration, and the changeset has an error
-  for the `:email` field with the message `has already been taken`, the user
-  will see the same success flow as above, but without any e-mail sent. This
-  will prevent information leak.
+  confirmed, a confirmation e-mail is sent, the session will be cleared, an
+  error flash is set for the `conn` and the user redirected back to
+  `Pow.Phoenix.Routes.after_registration_path/1` or
+  `Pow.Phoenix.Routes.after_sign_in_path/1` respectively.
 
   ### User updates e-mail
 
@@ -27,13 +19,22 @@ defmodule PowEmailConfirmation.Phoenix.ControllerCallbacks do
   `PowInvitation.Phoenix.InvitationController.update/2`
 
   When a user changes their e-mail, a confirmation e-mail is send to the new
-  e-mail, and an error flash is set for the conn. The same happens if the
+  e-mail, and an error flash is set for the `conn`. The same happens if the
   `PowInvitation` extension is enabled, and a user updates their e-mail when
   accepting their invitation. It's assumed that the current e-mail for the
   invited user has already been confirmed, see
   `PowInvitation.Ecto.Schema.invite_changeset/3` for more.
 
   See `PowEmailConfirmation.Ecto.Schema` for more.
+
+  ### Unique constraint error on `:email`
+
+  Triggers on `Pow.Phoenix.RegistrationController.create/2`.
+
+  When a user can't be created and the changeset has a unique constraint error
+  for the `:email` field, the user will experience the same success flow as if
+  the user could be created, but no e-mail is sent out. This prevents
+  information leak.
   """
   use Pow.Extension.Phoenix.ControllerCallbacks.Base
 
@@ -50,10 +51,10 @@ defmodule PowEmailConfirmation.Phoenix.ControllerCallbacks do
     halt_unconfirmed(conn, {:ok, user, conn}, return_path)
   end
   def before_respond(Pow.Phoenix.RegistrationController, :create, {:error, changeset, conn}, _config) do
-    case prevent_information_leak?(conn, changeset) do
+    case Plug.__prevent_information_leak__(conn, changeset) do
       true ->
         return_path = routes(conn).after_registration_path(conn)
-        conn        = redirect_and_show_error(conn, return_path)
+        conn        = redirect_with_email_confirmation_required(conn, return_path)
 
         {:halt, conn}
 
@@ -61,15 +62,17 @@ defmodule PowEmailConfirmation.Phoenix.ControllerCallbacks do
         {:error, changeset, conn}
     end
   end
+  def before_respond(Pow.Phoenix.RegistrationController, :update, {:ok, user, conn}, _config) do
+    warn_unconfirmed(conn, user)
+  end
   def before_respond(Pow.Phoenix.SessionController, :create, {:ok, conn}, _config) do
     return_path = routes(conn).after_sign_in_path(conn)
 
     halt_unconfirmed(conn, {:ok, conn}, return_path)
   end
-  def before_respond(Pow.Phoenix.RegistrationController, :update, {:ok, user, conn}, _config),
-    do: warn_unconfirmed(conn, user)
-  def before_respond(PowInvitation.Phoenix.InvitationController, :update, {:ok, user, conn}, _config),
-    do: warn_unconfirmed(conn, user)
+  def before_respond(PowInvitation.Phoenix.InvitationController, :update, {:ok, user, conn}, _config) do
+    warn_unconfirmed(conn, user)
+  end
 
   defp halt_unconfirmed(conn, success_response, return_path) do
     case PowEmailConfirmationPlug.email_unconfirmed?(conn) do
@@ -84,26 +87,18 @@ defmodule PowEmailConfirmation.Phoenix.ControllerCallbacks do
 
     send_confirmation_email(user, conn)
 
-    conn = redirect_and_show_error(conn, return_path)
+    conn = redirect_with_email_confirmation_required(conn, return_path)
 
     {:halt, conn}
   end
 
-  defp redirect_and_show_error(conn, return_path) do
+  defp redirect_with_email_confirmation_required(conn, return_path) do
     error = extension_messages(conn).email_confirmation_required(conn)
 
     conn
     |> Phoenix.Controller.put_flash(:error, error)
     |> Phoenix.Controller.redirect(to: return_path)
   end
-
-  defp prevent_information_leak?(_conn, %{errors: errors}) do
-    Enum.find_value(errors, false, fn
-      {:email, {_msg, [constraint: :unique, constraint_name: _name]}} -> true
-      _any                                                            -> false
-    end)
-  end
-  defp prevent_information_leak?(_conn, _changeset), do: false
 
   defp warn_unconfirmed(%{params: %{"user" => %{"email" => email}}} = conn, %{unconfirmed_email: email} = user) do
     case PowEmailConfirmationPlug.pending_email_change?(conn) do